% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvar_is_set.R
\name{envvar_is_set}
\alias{envvar_is_set}
\title{Check whether an environment variable is set}
\usage{
envvar_is_set(x)
}
\arguments{
\item{x}{String with the name of environment variable}
}
\value{
A logical value
}
\description{
\code{envvar_is_set()} checks whether a given environment variable is
set.
}
\examples{
envvar_is_set("HOME")
}
