% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvar_list.R
\name{envvar_list}
\alias{envvar_list}
\title{Get a list of environment variables}
\usage{
envvar_list()
}
\value{
A named list
}
\description{
\code{envvar_list()} returns a named list of the defined environment variables
}
\examples{
envvar_list()
}
