% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvar_set.R
\name{envvar_set}
\alias{envvar_set}
\alias{envvar_unset}
\title{Set, unset, and check environment variables}
\usage{
envvar_set(...)

envvar_unset(x)
}
\arguments{
\item{...}{Named arguments containing an environment variable and its value}

\item{x}{String with the name of environment variable}
}
\value{
No return value, called for side effects
}
\description{
\code{envvar_set()} sets one or more environment variables.

\code{envvar_unset()} unsets an environment variable.
}
\examples{
envvar_set(DEBUG = 1)
envvar_unset("DEBUG")
}
