% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{envvar_get_file}
\alias{envvar_get_file}
\alias{envvar_get_dir}
\title{Environment variables for files and directories}
\usage{
envvar_get_file(
  x,
  create = TRUE,
  check_readable = FALSE,
  check_writable = FALSE,
  transform = NULL,
  ...
)

envvar_get_dir(
  x,
  create = TRUE,
  transform = NULL,
  check_readable = FALSE,
  check_writable = FALSE,
  ...
)
}
\arguments{
\item{x}{String containing an environment variable name}

\item{create}{Create the file or directory if it does not exist (default:
\code{TRUE})}

\item{check_readable}{Ensure that the file or directory is readable}

\item{check_writable}{Ensure that the file or directory is writable}

\item{transform}{Optional function that applies a transformation to the
variable's value}

\item{...}{Additional arguments passed to \link[fs:create]{fs::file_create} for
\code{envvar_get_file()} or \link[fs:create]{fs::dir_create} for \code{envvar_get_dir()}}
}
\description{
\code{envvar_get_file} gets a file name from an environment variable

\code{envvar_get_dir} gets a directory name from an environment
variable
}
\examples{

# Get a file path and make sure it exists
\dontrun{
envvar_set("MYDATA" = "/tmp/data.parquet")
envvar_get_file("MYDATA")
}

# Get a file path and make sure it exists
\dontrun{
envvar_set("MYDATADIR" = "/tmp/data/")
envvar_get_dir("MYDATADIR")
}
}
