% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvar_get_list.R
\name{envvar_get_list}
\alias{envvar_get_list}
\title{Environment variables containing lists}
\usage{
envvar_get_list(
  x,
  pattern = ":",
  default = NA,
  validate = NULL,
  use_default = TRUE,
  ...
)
}
\arguments{
\item{x}{String containing an environment variable name}

\item{pattern}{String specifying the pattern used to separate elements in the
list.}

\item{default}{Optional default value if the environment variable is not set}

\item{validate}{Optional function that checks a value for validity}

\item{use_default}{Logical value indicating whether the value specified by
\code{default} should be used (\code{TRUE}, the default) or an error should be raised
(\code{FALSE})}

\item{...}{Additional arguments passed to \link{strsplit}}
}
\value{
A list
}
\description{
\code{envvar_get_list() } gets lists from environment variables. At
the moment, only unnamed lists are supported.
}
\examples{
# Get the value of `$PATH`, creating a list with elements for each directory
envvar_get_list("PATH")
}
