% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvar-package.R
\docType{package}
\name{envvar-package}
\alias{envvar}
\alias{envvar-package}
\title{envvar: Make Working with Environment Variables Easier and More Consistent}
\description{
A collection of functions that allows for easy and consistent use of environment variables. This includes setting, checking, retrieving, transforming, and validating values stored in environment variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/briandconnelly/envvar}
  \item \url{https://briandconnelly.github.io/envvar/}
  \item Report bugs at \url{https://github.com/briandconnelly/envvar/issues}
}

}
\author{
\strong{Maintainer}: Brian Connelly \email{bdc@bconnelly.net} (\href{https://orcid.org/0000-0002-9948-0379}{ORCID}) [copyright holder]

}
\keyword{internal}
