% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvar-package.R
\docType{package}
\name{envvar-package}
\alias{envvar}
\alias{envvar-package}
\title{envvar: Make working with environment variables easier and more consistent}
\description{
envvar makes it easy to set, check, retrieve, transform, and validate environment variables.
}
\author{
\strong{Maintainer}: Brian Connelly \email{bdc@bconnelly.net} (\href{https://orcid.org/0000-0002-9948-0379}{ORCID}) [copyright holder]

}
\keyword{internal}
